requirements.txt
Django==1.4.2
dj-database-url==0.2.1
django-allauth==0.8.2
django-social-auth==0.7.7
django-tastypie==0.9.12-alpha
#-e git+https://github.com/toastdriven/django-tastypie.git#egg=django-tastypie
httplib2==0.7.6
mimeparse==0.1.3
oauth2==1.5.211
psycopg2==2.4.5
python-dateutil==1.5
python-openid==2.2.5
wsgiref==0.1.2
